Fixed prototype.
Rev: src/modules/HTTPLoop/util.h:1.3

+++ src/modules/HTTPLoop/util.h
@@ -1,5 +1,5 @@
 int aap_get_time(void);
-int aap_swrite(int to, char *buf, size_t towrite);
+ptrdiff_t aap_swrite(int to, char *buf, size_t towrite);
 
 #define H_EXISTS 0
 #define H_INT    1
